+++ stdlib/snq.h
@@ -30,6 +30,9 @@ void queue_destroy(Queue *q);
 void queue_push(Queue *q, void *item, int elem_size);
 void *queue_pop(Queue *q);
 void *queue_peek(Queue *q);
+int queue_size(Queue *q);
+void queue_clear(Queue *q);
+int queue_isEmpty(Queue *q);
 
 // Stack functions (no threads)
 Stack stack_create(int capacity);
@@ -37,5 +40,8 @@ void stack_destroy(Stack *s);
 void stack_push(Stack *s, void *item, int elem_size);
 void *stack_pop(Stack *s);
 void *stack_peek(Stack *s);
+int stack_size(Stack *s);
+void stack_clear(Stack *s);
+int stack_isEmpty(Stack *s);
 
 #endif